Переносит функции для фильтрации задач в контроллер фильтра
