+++ src/controllers/board.js
@@ -10,7 +10,6 @@ import {SortType, BoardMode as Mode, TaskMode as TaskControllerMode} from "../co
 const SHOWING_TASKS_AMOUNT_ON_START = 8;
 const SHOWING_TASKS_AMOUNT_BY_BUTTON = 8;
 
-
 const renderTasks = (taskListElement, tasks, onDataChange, onViewChange) => {
   return tasks.map((task) => {
     const taskController = new TaskController(taskListElement, onDataChange, onViewChange);
